Refactor BulletLabel
