+++ SwiftUIWeeklyLayoutChallenge/Topic005.swift
@@ -76,25 +76,6 @@ public struct Topic005View: View {
     }
 }
 
-private struct BulletLabel: View {
-    private let title: LocalizedStringKey
-    private let mark: String
-    
-    init(_ title: LocalizedStringKey, mark: String = "•") {
-        self.title = title
-        self.mark = mark
-    }
-    
-    var body: some View {
-        Label {
-            Text(title)
-        } icon: {
-            Text(mark)
-        }
-        .labelStyle(.bullet)
-    }
-}
-
 struct Topic005View_Previews: PreviewProvider {
     static var previews: some View {
         Topic005View()
